requirements.txt
requests==2.26.0
starlette==0.16.0
aiogram==2.14.3
aiohttp==3.7.4.post0
pycryptodome==3.10.1
